refactor: LegacyBoardGame 정리
