--- src/game/phaserConfig.tsx
@@ -1,2 +1,0 @@
-﻿export { default } from '../components/board/LegacyBoardGame'
-export type { BoardGameHandle } from '../components/board/LegacyBoardGame'
