+++ src/hooks/game/useDiceRoll.ts
@@ -1,5 +1,5 @@
 import { RefObject, useCallback } from 'react'
-import type { BoardGameHandle } from '../../components/board/LegacyBoardGame'
+import type { BoardGameHandle } from '../../components/board/GameBoard'
 import { socket } from '../../lib/socket'
 import { emitRollDice } from '../../services/socket/game.handler'
 import { useGameStore } from '../../stores/game.store'
