+++ src/pages/GamePage.tsx
@@ -11,7 +11,7 @@ import { useGameState } from '../hooks/game/useGameState'
 import { useGameTimer } from '../hooks/game/useGameTimer'
 import { useDiceRoll } from '../hooks/game/useDiceRoll'
 import { useTurn } from '../hooks/game/useTurn'
-import BoardGame, { BoardGameHandle } from '../components/board/LegacyBoardGame'
+import BoardGame, { BoardGameHandle } from '../components/board/GameBoard'
 import { INIT_PLAYERS, PlayerState } from '../components/board/board.constants'
 import type { BuildingLevel, ChatMessage } from '../types/domain'
 import { socket } from '../lib/socket'
